+++ lib/src/ui/form/Form.dart
@@ -1,0 +1,10 @@
+import 'package:tch_common_widgets/src/ui/form/PreferencesSwitchWidget.dart';
+
+class FormStyle {
+  final PreferencesSwitchStyle preferencesSwitchStyle;
+
+  /// FormStyle initialization
+  const FormStyle({
+    this.preferencesSwitchStyle = const PreferencesSwitchStyle(),
+  });
+}