+++ lib/tch_common_widgets.dart
@@ -1,6 +1,8 @@
 library tch_common_widgets;
 
-export 'src/core/CommonTheme.dart';
 export 'src/core/CommonDimens.dart';
+export 'src/core/CommonTheme.dart';
 
+export 'src/ui/form/Form.dart';
+export 'src/ui/form/PreferencesSwitchWidget.dart';
 export 'src/ui/widgets/CommonSpace.dart';